add `queryhelp` property to QueryBuilder
 * add `queryhelp` property to QueryBuilder
   (currently pipes through to get_json_compatible_queryhelp)
 * add __deepcopy__ function to allow direct copies of QueryBuilder
   instances
 * test that QueryBuilder can be re-created from queryhelp
